Post-Publish Initial Commit, Gitignore moved
This initial commit includes all of the required initial project files associated with the project, as well as their folders. Cloning this repository will allow you to launch the project, where you can begin work.

Note: When working in this repository, consider doing work in a separate branch dedicated to the task you are doing, merging when complete.