Audio Added to Soldier, Camera Shake Added
Enemies now play audio when damaged. On-Death audio currently not functional, will be fixed in a later push. Camera Shake added, camera now shakes and returns to its original position whenever the player is damaged.

Added audio feedback to investigation to make it clearer when an enemy is investigating something versus seeing the player.